Small changes
Endpoints with documents related bussines logic, moved to document module and added some small things

+++ views/document.py
@@ -3,6 +3,7 @@ from fastapi.responses import JSONResponse, StreamingResponse
 
 from dotenv import load_dotenv
 import os
+import asyncio
 
 from views.auth import auth_requierd
 from db.db import check_permission, get_db
@@ -46,6 +47,7 @@ async def upload_s3_file(file: UploadFile, project_id: int):
             Body=contents,
             ContentType=file.content_type
         )
+    return file.filename
 
 async def check_file_extension(files: list[UploadFile]) -> bool:
     if not isinstance(files, list):
@@ -71,9 +73,38 @@ async def check_file_extension(files: list[UploadFile]) -> bool:
 
     return True
 
+@router.get("/projects/{project_id}/documents")
+async def get_project_documents(project_id: str = Path(...), user_payload: dict = Depends(auth_requierd)) -> JSONResponse:
+    user_id = user_payload["sub"]
+    project_id = int(project_id)
+    with get_db() as conn:
+        user_premission = check_permission(conn, user_id, project_id)
+    
+    if user_premission is not None:
+        response = await get_s3_documents_list(project_id)
+        return JSONResponse(response, status.HTTP_200_OK)
+    else:
+        raise HTTPException(status.HTTP_401_UNAUTHORIZED, "Unauthorized")
+
+@router.post("/projects/{project_id}/documents")
+async def upload_project_documents(files: list[UploadFile] = File(...), project_id: str = Path(...), user_payload: dict = Depends(auth_requierd)) -> JSONResponse:
+
+    await check_file_extension(files)
+    
+    with get_db() as conn:
+        user_permission = check_permission(conn, user_payload["sub"], project_id)
+    
+    if user_permission is None:
+        raise HTTPException(status.HTTP_401_UNAUTHORIZED, "Unauthorized")
+        
+
+    uploaded_files = await asyncio.gather(*(upload_s3_file(file, project_id) for file in files))
+    return JSONResponse(f"Files uploaded successfully: {uploaded_files}", status.HTTP_200_OK)
+
+
 # TODO: Exception for file not found
 @router.get("/projects/{project_id}/documents/{document_id}")
-async def get_s3_document(project_id: str, download_web: str = False , document_id: str = Path(...), user_payload: dict = Depends(auth_requierd)) -> None:
+async def download_project_document(project_id: str, download_web: str = False , document_id: str = Path(...), user_payload: dict = Depends(auth_requierd)) -> None:
     key = f"{project_id}/{document_id}"
     with get_db() as conn:
         user_perm = check_permission(conn, user_payload["sub"], int(project_id))
@@ -108,9 +139,9 @@ async def get_s3_document(project_id: str, download_web: str = False , document_
         raise HTTPException(status.HTTP_401_UNAUTHORIZED, "Unauthorized")
 
 
-@router.post("/projects/{project_id}/documents/{document_id}")
-async def update_s3_file(file: UploadFile = File(...),project_id: str = Path(...), document_id: str = Path(...), user_payload: dict = Depends(auth_requierd)):
-    key = f"{project_id}/{document_id}"
+@router.put("/projects/{project_id}/documents/{document_id}")
+async def update_porject_document(file: UploadFile = File(...), project_id: str = Path(...), document_id: str = Path(...), user_payload: dict = Depends(auth_requierd)):
+    key = f"{project_id}/{file.filename}"
 
     await check_file_extension([file])
 
@@ -136,9 +167,13 @@ async def update_s3_file(file: UploadFile = File(...),project_id: str = Path(...
             raise HTTPException(status_code=500, detail=str(e))
 
 @router.delete("/projects/{project_id}/documents/{document_id}")
-async def delete_s3_document(project_id: str = Path(...), document_id: str = Path(...)) -> JSONResponse:
+async def delete_project_document(project_id: str = Path(...), document_id: str = Path(...)) -> JSONResponse:
     key = f"{project_id}/{document_id}"
     try:
+        docs_list = await get_s3_documents_list(project_id=project_id)
+        if document_id not in docs_list:
+            raise HTTPException(status.HTTP_404_NOT_FOUND, "File doesn't exists")
+
         async with session.resource("s3") as s3:
             bucket = await s3.Bucket(BUCKET_NAME)
             await bucket.delete_objects(
@@ -149,6 +184,6 @@ async def delete_s3_document(project_id: str = Path(...), document_id: str = Pat
                     'Quiet': True
                 }
             )
-
     except ClientError as e:
         raise HTTPException(status.HTTP_400_BAD_REQUEST, detail=str(e))
+    return JSONResponse(content=None, status_code=status.HTTP_204_NO_CONTENT)
